requirements.txt
matplotlib == 3.1.2
networkit == 5.0.1
networkx == 2.4
snap == 0.5
cython == 0.29.14
numpy == 1.17.4
pandas == 0.25.3
pytest == 5.3.0
pillow == 7.0.0